extract server as a subproject
