+++ all/build.sbt
@@ -1,0 +1,2 @@
+name := "scala-repl-pp-all"
+
