+++ core/src/test/resources/demo-project/build.sbt
