+++ core/src/test/resources/demo-project/project/build.properties
