+++ core/src/test/resources/demo-project/project/plugins.sbt
