+++ core/src/test/resources/demo-project/src/main/scala/stringcalc/Main.scala
