+++ server/src/main/scala/replpp/server/EmbeddedRepl.scala
